Add files via upload
Uploaded 29 June 2024
Consolidates and simplifies  the various documents on CircuitSetup.s excellent 6-Channel ESP32 Energy Meter

THIS IS BASIC SOFTWARE. All it does is read and Serial.print out the power readings from the board and ESP32 in a formatted (sort of) table. 
The purpose is to get started with this board. MQTT and other enhancements are available elsewhere.  
I'm a Blynk.io fan, myself.